CORE-4690 Ensure subscriptions are closed before recreating on receiving new configuration in member services.
