+++ components/membership/membership-client-impl/src/test/kotlin/net/corda/membership/impl/client/MemberOpsClientTest.kt
@@ -14,9 +14,15 @@ import net.corda.data.membership.rpc.response.RegistrationStatus
 import net.corda.libs.configuration.SmartConfig
 import net.corda.lifecycle.LifecycleCoordinator
 import net.corda.lifecycle.LifecycleCoordinatorFactory
+import net.corda.lifecycle.LifecycleCoordinatorName
+import net.corda.lifecycle.LifecycleEventHandler
+import net.corda.lifecycle.LifecycleStatus
+import net.corda.lifecycle.RegistrationHandle
+import net.corda.lifecycle.RegistrationStatusChangeEvent
+import net.corda.lifecycle.StartEvent
+import net.corda.lifecycle.StopEvent
 import net.corda.membership.client.dto.MemberRegistrationRequestDto
 import net.corda.membership.client.dto.RegistrationActionDto
-import net.corda.membership.impl.client.lifecycle.MemberOpsClientLifecycleHandler
 import net.corda.messaging.api.exception.CordaRPCAPISenderException
 import net.corda.messaging.api.publisher.RPCSender
 import net.corda.messaging.api.publisher.factory.PublisherFactory
@@ -28,7 +34,10 @@ import org.junit.jupiter.api.Test
 import org.mockito.kotlin.any
 import org.mockito.kotlin.doAnswer
 import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
+import org.mockito.kotlin.never
+import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import java.time.Instant
 import java.util.concurrent.CompletableFuture
@@ -41,20 +50,22 @@ class MemberOpsClientTest {
     companion object {
         private const val HOLDING_IDENTITY_ID = "nodeId"
     }
+    private val componentHandle: RegistrationHandle = mock()
+    private val configHandle: AutoCloseable = mock()
 
     private var coordinatorIsRunning = false
     private val coordinator: LifecycleCoordinator = mock {
         on { isRunning } doAnswer { coordinatorIsRunning }
         on { start() } doAnswer { coordinatorIsRunning = true }
         on { stop() } doAnswer { coordinatorIsRunning = false }
+        on { followStatusChangesByName(any()) } doReturn componentHandle
     }
 
-    private var lifecycleHandler: MemberOpsClientLifecycleHandler? = null
+    private var lifecycleHandler: LifecycleEventHandler? = null
 
     private val lifecycleCoordinatorFactory: LifecycleCoordinatorFactory = mock {
-        on { createCoordinator(any(), any()) } doReturn coordinator
         on { createCoordinator(any(), any()) } doAnswer {
-            lifecycleHandler = it.arguments[1] as MemberOpsClientLifecycleHandler
+            lifecycleHandler = it.arguments[1] as LifecycleEventHandler
             coordinator
         }
     }
@@ -65,7 +76,9 @@ class MemberOpsClientTest {
 
     private lateinit var publisherFactory: PublisherFactory
 
-    private val configurationReadService: ConfigurationReadService = mock()
+    private val configurationReadService: ConfigurationReadService = mock {
+        on { registerComponentForUpdates(any(), any()) } doReturn configHandle
+    }
 
     private lateinit var memberOpsClient: MemberOpsClientImpl
 
@@ -79,15 +92,22 @@ class MemberOpsClientTest {
         ConfigKeys.MESSAGING_CONFIG to messagingConfig
     )
 
+    fun startComponent() = lifecycleHandler?.processEvent(StartEvent(), coordinator)
+    fun stopComponent() = lifecycleHandler?.processEvent(StopEvent(), coordinator)
+    fun changeRegistrationStatus(status: LifecycleStatus) = lifecycleHandler?.processEvent(
+        RegistrationStatusChangeEvent(mock(), status), coordinator
+    )
+    fun changeConfig() = lifecycleHandler?.processEvent(
+        ConfigChangedEvent(setOf(ConfigKeys.BOOT_CONFIG, ConfigKeys.MESSAGING_CONFIG), configs),
+        coordinator
+    )
+
     @Suppress("UNCHECKED_CAST")
     private fun setUpRpcSender() {
         // re-sets the rpc request
         rpcRequest = null
         // kicks off the MessagingConfigurationReceived event to be able to mock the rpc sender
-        lifecycleHandler?.processEvent(
-            ConfigChangedEvent(setOf(ConfigKeys.BOOT_CONFIG, ConfigKeys.MESSAGING_CONFIG), configs),
-            coordinator
-        )
+        changeConfig()
     }
 
     private val request = MemberRegistrationRequestDto(HOLDING_IDENTITY_ID, RegistrationActionDto.REQUEST_JOIN)
@@ -311,4 +331,111 @@ class MemberOpsClientTest {
         assertTrue { ex.message!!.contains("Expected class") }
         memberOpsClient.stop()
     }
+
+    @Test
+    fun `start event starts following the statuses of the required dependencies`() {
+        startComponent()
+
+        verify(coordinator).followStatusChangesByName(
+            eq(setOf(LifecycleCoordinatorName.forComponent<ConfigurationReadService>()))
+        )
+    }
+
+    @Test
+    fun `start event closes dependency handle if it exists`() {
+        startComponent()
+        startComponent()
+
+        verify(componentHandle).close()
+    }
+
+    @Test
+    fun `stop event doesn't closes handles before they are created`() {
+        stopComponent()
+
+        verify(componentHandle, never()).close()
+        verify(configHandle, never()).close()
+    }
+
+    @Test
+    fun `component handle is created after starting and closed when stopping`() {
+        startComponent()
+        stopComponent()
+
+        verify(componentHandle).close()
+    }
+
+    @Test
+    fun `config handle is created after registration status changes to UP and closed when stopping`() {
+        changeRegistrationStatus(LifecycleStatus.UP)
+        stopComponent()
+
+        verify(configHandle).close()
+    }
+
+    @Test
+    fun `registration status UP registers for config updates`() {
+        changeRegistrationStatus(LifecycleStatus.UP)
+
+        verify(configurationReadService).registerComponentForUpdates(
+            any(), any()
+        )
+        verify(coordinator, never()).updateStatus(eq(LifecycleStatus.UP), any())
+    }
+
+    @Test
+    fun `registration status DOWN sets component status to DOWN`() {
+        startComponent()
+        changeRegistrationStatus(LifecycleStatus.UP)
+        changeRegistrationStatus(LifecycleStatus.DOWN)
+
+        verify(configHandle).close()
+        verify(coordinator).updateStatus(eq(LifecycleStatus.DOWN), any())
+    }
+
+    @Test
+    fun `registration status ERROR sets component status to DOWN`() {
+        startComponent()
+        changeRegistrationStatus(LifecycleStatus.UP)
+        changeRegistrationStatus(LifecycleStatus.ERROR)
+
+        verify(configHandle).close()
+        verify(coordinator).updateStatus(eq(LifecycleStatus.DOWN), any())
+    }
+
+    @Test
+    fun `registration status DOWN closes config handle if status was previously UP`() {
+        startComponent()
+        changeRegistrationStatus(LifecycleStatus.UP)
+
+        verify(configurationReadService).registerComponentForUpdates(
+            any(), any()
+        )
+
+        changeRegistrationStatus(LifecycleStatus.DOWN)
+
+        verify(configHandle).close()
+    }
+
+    @Test
+    fun `registration status UP closes config handle if status was previously UP`() {
+        changeRegistrationStatus(LifecycleStatus.UP)
+
+        verify(configurationReadService).registerComponentForUpdates(
+            any(), any()
+        )
+
+        changeRegistrationStatus(LifecycleStatus.UP)
+
+        verify(configHandle).close()
+    }
+
+    @Test
+    fun `after receiving the messaging configuration the rpc sender is initialized`() {
+        changeConfig()
+        verify(publisherFactory).createRPCSender(any<RPCConfig<MembershipRpcRequest, MembershipRpcResponse>>(), any())
+        verify(rpcSender).start()
+        verify(coordinator).updateStatus(eq(LifecycleStatus.UP), any())
+    }
+
 }
