+++ components/membership/membership-group-read-impl/src/main/kotlin/net/corda/membership/impl/read/subscription/MembershipGroupReadSubscriptions.kt
@@ -83,9 +83,8 @@ interface MembershipGroupReadSubscriptions : Lifecycle {
                 config
             ).apply {
                 start()
-                val tempSub = memberListSubscription
+                memberListSubscription?.close()
                 memberListSubscription = this
-                tempSub?.close()
             }
         }
