+++ components/membership/membership-service-impl/src/main/kotlin/net/corda/membership/service/impl/MemberOpsServiceImpl.kt
@@ -129,7 +129,7 @@ class MemberOpsServiceImpl @Activate constructor(
         logger.info("Creating RPC subscription for '{}' topic", Schemas.Membership.MEMBERSHIP_RPC_TOPIC)
         val messagingConfig = event.config.toMessagingConfig()
         val processor = MemberOpsServiceProcessor(registrationProxy, virtualNodeInfoReadService)
-        val current = subscription
+        subscription?.close()
         subscription = subscriptionFactory.createRPCSubscription(
             rpcConfig = RPCConfig(
                 groupName = GROUP_NAME,
@@ -141,6 +141,5 @@ class MemberOpsServiceImpl @Activate constructor(
             responderProcessor = processor,
             nodeConfig = messagingConfig
         ).also { it.start() }
-        current?.close()
     }
 }
